Rajat develop

+++ postgres_database/database_blueprint.py
@@ -2,7 +2,7 @@ from flask import Blueprint, request
 import psycopg2
 from flask_restx import Api, Resource
 
-from util.utils import add_entry_to_db_repo, update_active_flag_db_repo, check_db_exists
+from util.utils import add_entry_to_db_repo, update_active_flag_db_repo, check_db_exists, get_vm_details
 
 database_blueprint = Blueprint("database_blueprint", __name__)
 database_api = Api(database_blueprint)
@@ -32,7 +32,7 @@ class CreateDatabase(Resource):
                 cursor.execute(create_user_query)
                 connection.commit()
         connection.close()
-        connection = psycopg2.connect("host='localhost' user=postgres password=test")
+        connection = psycopg2.connect("host='localhost' user=postgres password=postgres")
         database_creation_sql = "create database " + database_name + " owner " + user_name + ";"
         revoke_privileges_for_public = "revoke connect on database " + database_name + " from public"
         connection.autocommit = True
@@ -59,7 +59,9 @@ class ConnectDB(Resource):
         user_name = uname
         pwd = pwd
         try:
-            connection = psycopg2.connect(host='localhost', dbname=database_name, user=user_name, password=pwd)
+            db_vm = get_vm_details(database_name)
+            if db_vm is not None:
+                connection = psycopg2.connect(host=db_vm, dbname=database_name, user=user_name, password=pwd)
         except Exception as e:
             return "connection failed" + e.__str__()
         return "connection established"
@@ -72,16 +74,19 @@ class DropDatabase(Resource):
         # database_name = request.args.get("dbname")
         #database_name = dbname
         database_name = request.json['dbname']
-        connection = psycopg2.connect("host='localhost' user=postgres password=test")
-        connection.autocommit = True
-        database_deletion_sql = "drop database " + database_name + ";"
-        try:
-            with connection.cursor() as cursor:
-                cursor.execute(database_deletion_sql)
-                update_active_flag_db_repo(database_name, 'Inactive')
-                return "Database dropped Successfully"
-        except Exception as e:
-            return "Exception while dropping database : " + e.__str__()
+        db_vm = get_vm_details(database_name)
+        if db_vm is not None:
+            connection = psycopg2.connect("host='"+ db_vm + "' user=postgres password=test")
+            connection.autocommit = True
+            database_deletion_sql = "drop database " + database_name + ";"
+            try:
+                with connection.cursor() as cursor:
+                    cursor.execute(database_deletion_sql)
+                    update_active_flag_db_repo(database_name, 'Inactive')
+                    return "Database dropped Successfully"
+            except Exception as e:
+                return "Exception while dropping database : " + e.__str__()
+        return "No database found"
 
 
 @database_namespace.route("/dropDatabaseByOwner")
@@ -98,33 +103,36 @@ class DropDatabaseByOwner(Resource):
         res = verify_user(database_name, user_name, pwd)
         if res != "success":
             return "Can not verify user credentials to the database"
-        connection = psycopg2.connect("host='localhost' user=postgres password=test")
-        connection.autocommit = True
-        # collect db owner information
-        check_db_owner_query = 'SELECT d.datname as "Name", pg_catalog.pg_get_userbyid(d.datdba) as "Owner" FROM pg_catalog.pg_database d WHERE d.datname = ' + "'" + database_name + "'" + ' ORDER BY 1;'
-        owners = []
-        try:
-            with connection.cursor() as cursor:
-                cursor.execute(check_db_owner_query)
-                record = cursor.fetchall()
-                for n in record:
-                    owners.append(n)
-        except Exception as e:
-            return "Exception while getting owner information of a db: " + e.__str__()
+        db_vm = get_vm_details(database_name)
+        if db_vm is not None:
+            connection = psycopg2.connect("host='"+ db_vm +"' user=postgres password=postgres")
+            connection.autocommit = True
+            # collect db owner information
+            check_db_owner_query = 'SELECT d.datname as "Name", pg_catalog.pg_get_userbyid(d.datdba) as "Owner" FROM pg_catalog.pg_database d WHERE d.datname = ' + "'" + database_name + "'" + ' ORDER BY 1;'
+            owners = []
+            try:
+                with connection.cursor() as cursor:
+                    cursor.execute(check_db_owner_query)
+                    record = cursor.fetchall()
+                    for n in record:
+                        owners.append(n)
+            except Exception as e:
+                return "Exception while getting owner information of a db: " + e.__str__()
 
-        # if username match db owner, drop db
-        for row in owners:
-            owner = row[1]
-            if user_name == owner:
-                database_deletion_sql = "drop database " + database_name + ";"
-                try:
-                    with connection.cursor() as cursor:
-                        cursor.execute(database_deletion_sql)
-                        return "Database dropped Successfully"
-                except Exception as e:
-                    return "Exception while dropping database : " + e.__str__()
-        connection.close()
-        return "user is not the database owner"
+            # if username match db owner, drop db
+            for row in owners:
+                owner = row[1]
+                if user_name == owner:
+                    database_deletion_sql = "drop database " + database_name + ";"
+                    try:
+                        with connection.cursor() as cursor:
+                            cursor.execute(database_deletion_sql)
+                            return "Database dropped Successfully"
+                    except Exception as e:
+                        return "Exception while dropping database : " + e.__str__()
+            connection.close()
+            return "user is not the database owner"
+        return "No Database found"
 
 
 def verify_user(database_name, user_name, pwd):
@@ -165,16 +173,20 @@ class GetDatabaseSize(Resource):
         # url/getSize?dbname=<dbname>
         # database_name = request.args.get("dbname")
         database_name = dbname
-        connection = psycopg2.connect("host='localhost' user=postgres password=test")
-        database_size_sql = "SELECT pg_size_pretty( pg_database_size('" + database_name + "') );"
-        connection.autocommit = True
-        try:
-            with connection.cursor() as cursor:
-                cursor.execute(database_size_sql)
-                row = cursor.fetchone()
-                return str(row[0])
-        except Exception as e:
-            return "Exception while getting database size: " + e.__str__()
+        db_vm = get_vm_details(database_name)
+        if db_vm is not None:
+            connection = psycopg2.connect("host='" + db_vm + "' user=postgres password=postgres")
+            database_size_sql = "SELECT pg_size_pretty( pg_database_size('" + database_name + "') );"
+            connection.autocommit = True
+            try:
+                with connection.cursor() as cursor:
+                    cursor.execute(database_size_sql)
+                    row = cursor.fetchone()
+                    return str(row[0])
+            except Exception as e:
+                return "Exception while getting database size: " + e.__str__()
+        else:
+            return "Database Not found"
 
 
 @database_namespace.route("/getStats/<string:dbname>")
@@ -183,16 +195,19 @@ class GetDatabaseStats(Resource):
         # url/getStats?dbname=<dbname>
         # database_name = request.args.get("dbname")
         database_name = dbname
-        connection = psycopg2.connect("host='localhost' user=postgres password=test")
-        database_size_sql = "SELECT * FROM pg_stat_database WHERE datname = '" + database_name + "';"
-        connection.autocommit = True
-        try:
-            with connection.cursor() as cursor:
-                cursor.execute(database_size_sql)
-                row = cursor.fetchall()
-                return str(row)
-        except Exception as e:
-            return "Exception while getting database size: " + e.__str__()
+        db_vm = get_vm_details(database_name)
+        if db_vm is not None:
+            connection = psycopg2.connect("host='"+db_vm+"' user=postgres password=postgres")
+            database_size_sql = "SELECT * FROM pg_stat_database WHERE datname = '" + database_name + "';"
+            connection.autocommit = True
+            try:
+                with connection.cursor() as cursor:
+                    cursor.execute(database_size_sql)
+                    row = cursor.fetchall()
+                    return str(row)
+            except Exception as e:
+                return "Exception while getting database size: " + e.__str__()
+        return "Database not found"
 
 
 @database_namespace.route("/updateReadStatus")
@@ -204,25 +219,28 @@ class UpdateReadAccess(Resource):
         # user_name = request.args.get("uname")
         database_name = request.json['dbname']
         user_name = request.json['uname']
-        connection = psycopg2.connect("host='localhost' user=postgres password=test")
-        check_user_exists_query = "select 1 from pg_roles where pg_roles.rolname='" + user_name + "';"
-        with connection.cursor() as cursor:
-            cursor.execute(check_user_exists_query)
-            record = cursor.fetchall()
-            print(record)
-            if not record:
-                create_user_query = "create user " + user_name + " with password " + "'dummy_pwd#1234';"
-                cursor.execute(create_user_query)
-                connection.commit()
-            connection.close()
-            connection = psycopg2.connect("host='localhost' user=postgres password=test")
-            grant_connect_and_read_to_user = "grant connect on database " + database_name + " to " + user_name + ";"
-            # connection.autocommit = True
-            try:
-                with connection.cursor() as cursor:
-                    cursor.execute(grant_connect_and_read_to_user)
+        db_vm = get_vm_details(database_name)
+        if db_vm is not None:
+            connection = psycopg2.connect("host='"+db_vm+"' user=postgres password=test")
+            check_user_exists_query = "select 1 from pg_roles where pg_roles.rolname='" + user_name + "';"
+            with connection.cursor() as cursor:
+                cursor.execute(check_user_exists_query)
+                record = cursor.fetchall()
+                print(record)
+                if not record:
+                    create_user_query = "create user " + user_name + " with password " + "'dummy_pwd#1234';"
+                    cursor.execute(create_user_query)
                     connection.commit()
-                    return {"host": "128.31.27.249", "port": 5432, "database": database_name, "username": user_name,
-                            "password": "dummy_pwd#1234"}
-            except Exception as e:
-                return "Exception while creating database : " + e.__str__()
+                connection.close()
+                connection = psycopg2.connect("host='localhost' user=postgres password=postgres")
+                grant_connect_and_read_to_user = "grant connect on database " + database_name + " to " + user_name + ";"
+                # connection.autocommit = True
+                try:
+                    with connection.cursor() as cursor:
+                        cursor.execute(grant_connect_and_read_to_user)
+                        connection.commit()
+                        return {"host": "128.31.27.249", "port": 5432, "database": database_name, "username": user_name,
+                                "password": "dummy_pwd#1234"}
+                except Exception as e:
+                    return "Exception while creating database : " + e.__str__()
+        return "Database not found"
